# Add UniqueIds for improved scalable UUID generation