--- a/changelog/@unreleased/pr-1789.v2.yml
+++ b/changelog/@unreleased/pr-1789.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Add UniqueIds for improved scalable UUID generation
+  links:
+  - https://github.com/palantir/tritium/pull/1789