--- a/tritium-ids/src/test/java/com/palantir/tritium/ids/UniqueIdsTest.java
+++ b/tritium-ids/src/test/java/com/palantir/tritium/ids/UniqueIdsTest.java
@@ -41,7 +41,7 @@ public final class UniqueIdsTest {
 
     @Test
     public void deterministicRandomSource() {
-        UUID uuid = UniqueIds.v4RandomUuid(random);
+        UUID uuid = UniqueIds.randomUuidV4(random);
         assertThat(uuid).isEqualTo(UUID.fromString("ea6addb7-2596-428e-8a77-acf9a43797b3"));
         assertThat(uuid.variant()).isEqualTo(2); //  IETF RFC 4122
         assertThat(uuid.version()).isEqualTo(4); // v4 Randomly generated UUID
@@ -55,50 +55,50 @@ public final class UniqueIdsTest {
         assertThat(uuid.getLeastSignificantBits()).isEqualTo(0x8a77acf9a43797b3L);
         assertThat(uuid).isEqualTo(UUID.fromString(uuid.toString()));
 
-        assertThat(UniqueIds.v4RandomUuid(random))
+        assertThat(UniqueIds.randomUuidV4(random))
                 .isEqualTo(UUID.fromString("4b58fcf1-a036-4cf3-9ec0-4f2620d015d0"))
                 .extracting(UUID::version)
                 .isEqualTo(4);
-        assertThat(UniqueIds.v4RandomUuid(random))
+        assertThat(UniqueIds.randomUuidV4(random))
                 .isEqualTo(UUID.fromString("174cafe6-1614-44fb-b3af-b63e0b344571"))
                 .extracting(UUID::version)
                 .isEqualTo(4);
     }
 
     @ParameterizedTest
-    @ValueSource(ints = {1, 100, 1_000, 100_000})
+    @ValueSource(ints = {1, 1_000, 100_000, 256 * 1024})
     void concurrentRequests(int size) {
         ExecutorService executor = ForkJoinPool.commonPool();
         try {
             List<Future<UUID>> futures = new ArrayList<>(size);
             for (int i = 0; i < size; i++) {
-                futures.add(executor.submit(() -> UniqueIds.v4RandomUuid()));
+                futures.add(executor.submit(() -> UniqueIds.randomUuidV4()));
             }
             executor.shutdown();
             Iterator<Future<UUID>> iterator = futures.iterator();
-            assertRfc4122v4(() -> Futures.getUnchecked(iterator.next()), size);
+            assertRfc4122UuidV4(() -> Futures.getUnchecked(iterator.next()), size);
         } finally {
             executor.shutdownNow();
         }
     }
 
     @ParameterizedTest
-    @ValueSource(ints = {1, 100, 1_000, 100_000})
-    void randomUuid(int size) {
-        assertRfc4122v4(UniqueIds::v4RandomUuid, size);
+    @ValueSource(ints = {1, 1_000, 100_000, 256 * 1024})
+    void randomUuidV4(int size) {
+        assertRfc4122UuidV4(UniqueIds::randomUuidV4, size);
     }
 
-    private static void assertRfc4122v4(Supplier<UUID> supplier, int size) {
+    private static void assertRfc4122UuidV4(Supplier<UUID> supplier, int size) {
         Set<UUID> uuids = Sets.newHashSetWithExpectedSize(size);
         for (int i = 0; i < size; i++) {
             UUID uuid = supplier.get();
-            assertRfc4122v4(uuid);
+            assertRfc4122UuidV4(uuid);
             assertThat(uuids.add(uuid)).as("should be unique: %s", uuid).isTrue();
         }
         assertThat(uuids).hasSize(size);
     }
 
-    private static void assertRfc4122v4(UUID uuid) {
+    private static void assertRfc4122UuidV4(UUID uuid) {
         assertThat(uuid)
                 .extracting(UUID::variant)
                 .as("'%s' should be IETF RFC 4122", uuid)