--- a/versions.props
+++ b/versions.props
@@ -1,5 +1,6 @@
 com.fasterxml.jackson.*:* = 2.13.3
 com.fasterxml.jackson.core:* = 2.15.2
+com.fasterxml.uuid:* = 4.2.0
 com.github.ben-manes.caffeine:* = 3.1.8
 com.google.auto.service:* = 1.1.1
 com.google.code.findbugs:* = 3.0.2