## Before this PR

`UUID.randomUUID()` generation and string serialization can be a scalability bottleneck, especially under high volume AtlasDB workloads. The top two I have seen recently pop up in profiles:

- `com.palantir.atlasdb.transaction.impl.TransactionRetryStrategy.runWithRetry(IntPredicate, TransactionRetryStrategy$Retryable):71`
  - https://github.com/palantir/atlasdb/blob/1acfd3d41b3e7e504320ea76b40ee9c134dab91a/atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/TransactionRetryStrategy.java#L71
- `com.palantir.lock.client.LeasedLockToken.of(ConjureLockToken, Lease):39`
  - https://github.com/palantir/atlasdb/blob/1acfd3d41b3e7e504320ea76b40ee9c134dab91a/lock-api/src/main/java/com/palantir/lock/client/LeasedLockToken.java#L39

![image](https://github.com/palantir/atlasdb/assets/54594/394ab496-5b8c-4624-bbb3-101633c8f1c3)

There is some OpenJDK work either completed or in flight that is worth tracking:

- [JDK-8308804 Improve UUID.randomUUID performance with bulk/scalable PRNG access](https://bugs.openjdk.org/browse/JDK-8308804) , [PR 14135](https://github.com/openjdk/jdk/pull/14135) (not yet integrated, closed as stale)
- [JDK-8311207: Cleanup for Optimization for UUID.toString](https://bugs.openjdk.org/browse/JDK-8311207) , [PR 14745](https://github.com/openjdk/jdk/pull/14745)
- [JDK-8310502 : Optimization for j.l.Long.fastUUID()](https://bugs.openjdk.org/browse/JDK-8310502) , [PR 14578](https://github.com/openjdk/jdk/pull/14578)

Also worth tracking and possibly utilizing one of the newer proposed UUID formats:

- https://uuid6.github.io/uuid6-ietf-draft/
- https://datatracker.ietf.org/doc/draft-peabody-dispatch-new-uuid-format/
- https://www.ietf.org/archive/id/draft-peabody-dispatch-new-uuid-format-04.html#v8


## After this PR
Moving reusable UniqueIds from elsewhere to `tritium-ids` for use in  https://github.com/palantir/atlasdb/pull/6726 and other places.

==COMMIT_MSG==
Add UniqueIds for improved scalable UUID generation
==COMMIT_MSG==


```
16x Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz cores
VM version: JDK 17.0.8, OpenJDK 64-Bit Server VM, 17.0.8+7-LTS
Benchmark                          Mode  Cnt   Score   Error   Units
UuidBenchmark.jdkRandomUuidMax    thrpt    3   1.687 ? 1.487  ops/us
UuidBenchmark.jdkRandomUuidOne    thrpt    3   2.529 ? 0.081  ops/us
UuidBenchmark.jugSecureRandomMax  thrpt    3   1.976 ? 1.289  ops/us
UuidBenchmark.randomUuidV4Max     thrpt    3  27.644 ? 4.933  ops/us
UuidBenchmark.randomUuidV4One     thrpt    3   3.355 ? 0.308  ops/us
UuidBenchmark.secureRandomMax     thrpt    3   2.028 ? 1.458  ops/us
UuidBenchmark.threadLocalMax      thrpt    3   1.649 ? 1.633  ops/us
UuidBenchmark.threadLocalSha1Max  thrpt    3  27.464 ? 6.074  ops/us
UuidBenchmark.twoLongsMax         thrpt    3   0.699 ? 0.925  ops/us
```

```
MacBook Pro 2021 10x aarch64 cores
VM version: JDK 17.0.8.1, OpenJDK 64-Bit Server VM, 17.0.8.1+8-LTS
Benchmark                          Mode  Cnt   Score   Error   Units
UuidBenchmark.jdkRandomUuidMax    thrpt    3   2.261 ? 0.255  ops/us
UuidBenchmark.jdkRandomUuidOne    thrpt    3   3.913 ? 0.946  ops/us
UuidBenchmark.jugSecureRandomMax  thrpt    3   2.801 ? 1.441  ops/us
UuidBenchmark.randomUuidV4Max     thrpt    3  37.058 ? 6.880  ops/us
UuidBenchmark.randomUuidV4One     thrpt    3   4.650 ? 0.193  ops/us
UuidBenchmark.secureRandomMax     thrpt    3   2.865 ? 0.795  ops/us
UuidBenchmark.threadLocalMax      thrpt    3   2.263 ? 0.343  ops/us
UuidBenchmark.threadLocalSha1Max  thrpt    3  34.557 ? 8.399  ops/us
UuidBenchmark.twoLongsMax         thrpt    3   1.127 ? 0.302  ops/us
```


## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->