@svc-autorelease 2023-09-06
Released [0.76.0](https://github.com/palantir/tritium/releases/tag/0.76.0)